refactor openai settings
