+++ src/lib/components/LLMProviderList.svelte
@@ -29,5 +29,23 @@
     <Button on:click={handleNewProvider} variant="outline" class="border border-zinc-700"
       >Add custom LLM provider</Button
     >
+    <p class="leading-tight">
+      <small>
+        You can use Prompta with any 3rd party tools that support the OpenAI API such as Helicone or
+        local LLMs that expose OpenAI compatible APIs like <a
+          href="https://github.com/mudler/LocalAI"
+          class="text-blue-200 hover:underline"
+          target="_blank"
+          rel="noopener noreferrer">LocalAI</a
+        >
+        or
+        <a
+          href="https://github.com/BerriAI/litellm"
+          class="text-blue-200 hover:underline"
+          target="_blank"
+          rel="noopener noreferrer">LiteLLM</a
+        >.
+      </small>
+    </p>
   {/if}
 </div>
